Move result struct in result mod.
